mnist: Make mnist type float32

+++ datasets/mnist.py
@@ -13,8 +13,8 @@ class MNIST:
     def __init__(self):
         mnist = tf.keras.datasets.mnist
         (x_train, y_train), (x_test, y_test) = mnist.load_data()
-        self.x_train = x_train[..., None] / 127.5 - 1.
-        self.x_test = x_test[..., None] / 127.5 - 1.
+        self.x_train = x_train[..., None].astype(np.float32) / 127.5 - 1.
+        self.x_test = x_test[..., None].astype(np.float32) / 127.5 - 1.
         self.y_train, self.y_test = y_train, y_test
 
     def rawdata(self, train=True):
